+++ src/test/java/PlayerTest.java
@@ -1,8 +1,7 @@
-package test;
-
 import org.example.Player;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PlayerTest {
     @Test
